+++ pysftp.py
@@ -390,11 +390,12 @@ class Connection(object):
         self._sftp_connect()
         return self._sftp.open(remote_file, mode=mode, bufsize=bufsize)
 
-    def get_ciphers(self):
-        """Get list of currently used ciphers.
-        Ciphers are applied in order of first found on server.
+    @property
+    def active_ciphers(self):
+        """Get tuple of currently used local and remote ciphers.
 
         :returns: a tuple of currently used ciphers (local, remote)
+
         """
         return self._transport.local_cipher, self._transport.remote_cipher