+++ docs/changes.rst
@@ -5,8 +5,8 @@ Change Log
 * 0.2.5 (dev)
 
   * added ciphers parameter to Connection object
-  * added .get_ciphers method to return local and remote cipher in use
-  * added .security_options
+  * added .active_ciphers method to return local and remote cipher in use
+  * added .security_options, from where you can get available ciphers, among other information
 
 * 0.2.4 (current, released 2014-05-13)